% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello()
}
\arguments{
\item{NONE}{description}
}
\value{
None
}
\description{
A happy greeting made for RDSA
}
\section{This is an example function named 'hello'}{
which prints 'Hello, world!'.
}

\examples{
hello()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
