% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_untested.R
\name{hello_untested}
\alias{hello_untested}
\title{hello_untested}
\usage{
hello_untested()
}
\arguments{
\item{NONE}{description}
}
\value{
None
}
\description{
A happy greeting made for RDSA, BUT UNTESTED
}
\section{This is an example function named 'hello'}{
which prints 'Hello, world!'.
}

\examples{
hello_untested()
}
\author{
Mark London \email{mark.london@ext.ons.gov.uk}
}
