% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz.R
\docType{package}
\name{xyz}
\alias{xyz}
\alias{xyz-package}
\title{xyz: A package for fast interaction search in high dimensional data using the xyz algorithm.}
\description{
xyz: A package for fast interaction search in high dimensional data using the xyz algorithm.
}
\section{xyz functions}{

xyz_search,
xyz_regression
}

\references{
G. Thanei, N. Meinshausen and R. Shah (2016). The xyz algorithm for fast interaction search in high-dimensional data. <https://arxiv.org/pdf/1610.05108v1.pdf>
}
